+++ src/SegmentedToggle/ToggleIcon/ToggleIcon.st.css
@@ -9,7 +9,7 @@
 }
 
 :import {
-  -st-from: "../../Tooltip/TooltipNext/Tooltip.st.css";
+  -st-from: "../../Tooltip/Tooltip.st.css";
   -st-default: Tooltip;
 }
