+++ src/Tooltip/Tooltip.meta.js
@@ -4,5 +4,5 @@ import Registry from '@ui-autotools/registry';
 const metadata = Registry.getComponentMetadata(Tooltip);
 
 metadata.exportedFrom({
-  path: 'src/Tooltip/TooltipNext/Tooltip.js',
+  path: 'src/Tooltip',
 });
