--- src/Tooltip/TooltipContent.scss
@@ -1,219 +1,0 @@
-@import './common.scss';
-@import '../common.scss';
-@import '../ZIndex.scss';
-
-.root {
-  position: absolute;
-  z-index: $zIndex_TOOLTIP;
-  top: 0;
-  left: -9000px;
-}
-
-@mixin arrow($side, $color) {
-  @if $side == 'right' {
-    border-top: 5px solid $color;
-    border-left: 5px solid transparent;
-    border-bottom: 5px solid transparent;
-    border-right: 5px solid $color;
-  }
-  @if $side == 'left' {
-    border-top: 5px solid transparent;
-    border-left: 5px solid $color;
-    border-bottom: 5px solid $color;
-    border-right: 5px solid transparent;
-  }
-  @if $side == 'top' {
-    border-top: 5px solid $color;
-    border-left: 5px solid $color;
-    border-bottom: 5px solid transparent;
-    border-right: 5px solid transparent;
-  }
-  @if $side == 'bottom' {
-    border-bottom: 5px solid $color;
-    border-right: 5px solid $color;
-    border-top: 5px solid transparent;
-    border-left: 5px solid transparent;
-  }
-}
-
-.tooltip {
-  @include FontRoman;
-  @include popoverTheme;
-
-  position: relative;
-  display: inline-block;
-  padding: 12px 24px;
-  box-sizing: border-box;
-  font-size: 14px;
-  line-height: 1.29;
-
-  word-wrap: break-word;
-  -webkit-font-smoothing: antialiased;
-
-  &.light {
-    color: $D10;
-    background: $D80;
-  }
-
-  &.dark {
-    color: $D80;
-    background: $D10;
-  }
-
-  &.error {
-    color: white;
-    background: $R00;
-  }
-
-  &.normal {
-    font-size: 14px;
-    line-height: 18px;
-    text-align: center;
-  }
-
-  &.large {
-    font-size: 16px;
-    line-height: 24px;
-    text-align: center;
-  }
-}
-
-.popover-tooltip-content {
-  overflow: hidden;
-  word-wrap: normal;
-  vertical-align: top;
-  padding: 0;
-}
-
-.arrow {
-  $arrow-size: 10px;
-  $shadow-color-arrow: rgba(0, 0, 0, .1);
-  $arrowBottomShadow: 3px 3px 6px $shadow-color-arrow;
-  $arrowTopShadow: -3px -3px 6px $shadow-color-arrow;
-  $arrowRightShadow: 3px -3px 6px $shadow-color-arrow;
-  $arrowLeftShadow: -3px 3px 6px $shadow-color-arrow;
-
-  position: absolute;
-  height: $arrow-size;
-  width: $arrow-size;
-  transform: rotateZ(45deg);
-  background: inherit;
-
-  box-sizing: border-box;
-  background: transparent;
-
-
-  &.light.top {
-    @include arrow('top', $D80)
-  }
-  &.light.bottom {
-    @include arrow('bottom', $D80)
-  }
-  &.light.right {
-    @include arrow('right', $D80)
-  }
-  &.light.left {
-    @include arrow('left', $D80)
-  }
-
-  &.dark.top {
-    @include arrow('top', $D10)
-  }
-  &.dark.bottom {
-    @include arrow('bottom', $D10)
-  }
-  &.dark.right {
-    @include arrow('right', $D10)
-  }
-  &.dark.left {
-    @include arrow('left', $D10)
-  }
-
-  &.error.top {
-    @include arrow('top', $R00)
-  }
-  &.error.bottom {
-    @include arrow('bottom', $R00)
-  }
-  &.error.right {
-    @include arrow('right', $R00)
-  }
-  &.error.left {
-    @include arrow('left', $R00)
-  }
-
-  &.bottom,
-  &.top {
-    left: 50%;
-    margin-left: -$arrow-size/2;
-  }
-
-  &.bottom {
-    bottom: -5px;
-    box-shadow: $arrowBottomShadow;
-  }
-
-  &.top {
-    top: -5px;
-    box-shadow: $arrowTopShadow;
-  }
-
-  &.left,
-  &.right {
-    top: 50%;
-    margin-top: -5px;
-  }
-
-  &.right {
-    right: -5px;
-    box-shadow: $arrowRightShadow;
-  }
-
-  &.left {
-    left: -5px;
-    box-shadow: $arrowLeftShadow;
-  }
-}
-
-.fadeIn { animation: fadeIn .3s cubic-bezier(0.23, 1, 0.32, 1) both; }
-
-.bounce-top,
-.bounce-bottom { animation: bounceVertically 1s ease infinite; }
-
-.bounce-left,
-.bounce-right { animation: bounceHorizontally 1s ease infinite; }
-
-.bounce-top,
-.bounce-bottom,
-.bounce-left,
-.bounce-right {
-  &:hover {
-    animation-play-state: paused;
-    -webkit-animation-play-state: paused;
-    -moz-animation-play-state: paused;
-    -o-animation-play-state: paused;
-  }
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: .85;
-    transform: scale(.9);
-  }
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-@keyframes bounceVertically {
-    0% { transform: translateY(0); }
-   50% { transform: translateY(-10px); }
-  100% { transform: translateY(0); }
-}
-
-@keyframes bounceHorizontally {
-    0% { transform: translateX(0); }
-   50% { transform: translateX(-10px); }
-  100% { transform: translateX(0); }
-}
