--- src/Tooltip/TooltipNext/index.js
@@ -1,1 +1,0 @@
-export { default } from './Tooltip.js';
