--- src/Tooltip/docs/Composite/Example.scss
@@ -1,58 +1,0 @@
-@import '../../../common.scss';
-
-.option {
-  margin-bottom: 10px;
-}
-
-.flex {
-  display: flex;
-  align-items: center;
-  margin-top: 10px;
-}
-
-.form {
-  display: flex;
-  align-items: top;
-  padding: 20px;
-}
-
-.input { width: 60%; }
-.output {
-  width: 40%;
-  background-color: $GR30;
-}
-
-.output-lightblue {
-  background-color: $D70;
-  width: 35%;
-}
-
-.output-darkblue {
-  background-color: $D10;
-  width: 35%;
-}
-
-.output-white {
-  background-color: $D80;
-  width: 35%;
-}
-
-.example-wrapper {
-  padding: 80px;
-  text-align: center;
-}
-
-.root {
-  padding: 0 200px;
-}
-
-.box {
-  box-sizing: border-box;
-  display: flex;
-  align-items: center;
-  width: 200px;
-  min-height: 80px;
-  padding: 10px;
-  border: solid 1px #efefef;
-}
-
