--- src/Tooltip/docs/Core/Example.scss
@@ -1,12 +1,0 @@
-@import '../../../common.scss';
-
-.box {
-  box-sizing: border-box;
-  display: flex;
-  align-items: center;
-  width: 200px;
-  min-height: 80px;
-  padding: 10px;
-  border: solid 1px #efefef;
-}
-
