--- src/Tooltip/docs/Core/ExamplePlacement.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import Tooltip from 'wix-style-react/Tooltip';
-
-import styles from './Example.scss';
-
-export default () => (
-  <Tooltip
-    active
-    placement="right"
-    alignment="center"
-    content="Right Center"
-    showTrigger="custom"
-    hideTrigger="custom"
-  >
-    <div className={styles.box}>Right Center</div>
-  </Tooltip>
-);
