--- src/Tooltip/docs/content.js
@@ -1,3 +1,0 @@
-export const SHORT_CONTENT = 'Hello World';
-export const LONG_CONTENT =
-  'Hello World, Hello World, Hello World, Hello World';
