+++ src/Tooltip/test/storySettings.js
@@ -1,4 +1,4 @@
-import { Category } from '../../../../stories/storiesHierarchy';
+import { Category } from '../../../stories/storiesHierarchy';
 
 export const storySettings = {
   category: Category.COMPONENTS,
