+++ src/common/EllipsisHOC/EllipsisHOC.st.css
@@ -4,7 +4,7 @@
 }
 
 :import {
-  -st-from: "../../Tooltip/TooltipNext/Tooltip.st.css";
+  -st-from: "../../Tooltip/Tooltip.st.css";
   -st-default: Tooltip;
 }
